OAuth 2

+++ project/routers/reviews.py
@@ -14,6 +14,9 @@ from ..schemas import ReviewRequestPutModel
 from typing import List
 
 
+from fastapi import Depends
+from ..common import get_current_user
+
 
 router = APIRouter(prefix='/reviews')
 
@@ -22,19 +25,19 @@ router = APIRouter(prefix='/reviews')
 # ---------- Crear Reseña --------------
 
 @router.post('', response_model = ReviewResponseModel)
-async def create_reviews(user_review: ReviewRequestModel):
+async def create_reviews(user_review: ReviewRequestModel, user: User = Depends(get_current_user)):
 
     # validando si existe el usuario
-    if User.select().where(User.id == user_review.user_id).first() is None:
+    """ if User.select().where(User.id == user_review.user_id).first() is None:
         raise HTTPException(status_code = 404, detail = 'User not found')
-    
+     """
     # validando si existe la pelicula
     if Movie.select().where(Movie.id == user_review.movie_id).first() is None:
         raise HTTPException(status_code = 404, detail = 'Movie not found')
     
     
     user_review = UserReview.create(
-        user_id = user_review.user_id,
+        user_id = user.id, #user_review.user_id,
         movie_id = user_review.movie_id,
         reviews = user_review.reviews,
         score = user_review.score
@@ -71,12 +74,15 @@ async def get_review(id: int):
 # actualizar reseña
 
 @router.put('/{id}', response_model =  ReviewResponseModel)
-async def update_review(id: int, review_request: ReviewRequestPutModel):
+async def update_review(id: int, review_request: ReviewRequestPutModel, user: User = Depends(get_current_user)):
     review = UserReview.select().where(UserReview.id == id).first()
 
     # no existe
     if review is None:
         raise HTTPException(status_code = 404, detail = 'Review not found')
+
+    if review.user_id != user.id: #user.id -> el user autenticado
+        raise HTTPException(status_code = 401, detail = 'No eres el propietario')
 
     # si existe
     review.reviews = review_request.reviews
@@ -90,12 +96,16 @@ async def update_review(id: int, review_request: ReviewRequestPutModel):
 # eliminar reseña
 
 @router.delete('/{id}', response_model = ReviewResponseModel)
-async def delete_review(id: int):
+async def delete_review(id: int, user: User = Depends(get_current_user)):
     review = UserReview.select().where(UserReview.id == id).first()
 
     # no existe
     if review is None:
         raise HTTPException(status_code = 404, detail = 'Review not found')
+
+    if review.user_id != user.id: #user.id -> el user autenticado
+        raise HTTPException(status_code = 401, detail = 'No eres el propietario')
+
 
     # si existe
     review.delete_instance() # de peewee
